Move security policy building into API module so that it can be more easily reused outside the CLI verb
